refactor: 패키지명 변경 및 일부 메소드 수정 및 추가
