+++ src/main/java/com/practice/blog/base/config/SecurityConfig.java
@@ -1,0 +1,2 @@
+package com.practice.blog.base.config;public class SecurityConfig {
+}
