+++ src/main/java/com/practice/blog/bounded_context/auth/controller/AuthController.java
@@ -1,0 +1,2 @@
+package com.practice.blog.bounded_context.auth.controller;public class AuthController {
+}
