--- src/main/java/com/practice/blog/bounded_context/user/service/UserService.java
@@ -1,22 +1,0 @@
-package com.practice.blog.bounded_context.user.service;
-
-import com.practice.blog.bounded_context.user.entity.User;
-import com.practice.blog.bounded_context.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserService {
-
-    private final UserRepository userRepository;
-
-    public User read(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
-
-        return user;
-    }
-
-
-}
